% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD16S.R
\name{IBD16S}
\alias{IBD16S}
\title{Create HMP2 16S sequencing data for Inflammatory Bowel Disease (IBD) phyloseq object}
\format{A phyloseq object with 982 taxa and 178 samples.
\subsection{OTU Table}{
    A counts matrix for the rRNA-seq results.
}
\subsection{Sample Data}{
   \describe{
      \item{Project}{Name of the project}
      \item{sample_id}{Sample identifier}
      \item{subject_id}{Subject identifier}
      See colnames(sample_data(IBD16S())) for remaining values.
   }
}
\subsection{Taxonomy Table}{
    Taxonomy table for the taxa identified.
}}
\usage{
IBD16S()
}
\value{
a phyloseq object
}
\description{
The HMP2 study gathered data on the microbiome for
patients with different disease. This dataset contains
16S rRNA data for 196 patients with IBD over multiple visits.
}
\examples{
IBD <- IBD16S()
}
