% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD16S_mtx.R
\docType{data}
\name{IBD16S_mtx}
\alias{IBD16S_mtx}
\title{IBD16S_mtx}
\format{An object of class \code{matrix} with 982 rows and 178 columns.}
\usage{
IBD16S_mtx
}
\description{
Matrix of OTU values for 16S rRNA data for IBD
}
\keyword{datasets}
