% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD16S_samp.R
\docType{data}
\name{IBD16S_samp}
\alias{IBD16S_samp}
\title{IBD16S_samp}
\format{An object of class \code{data.frame} with 178 rows and 490 columns.}
\usage{
IBD16S_samp
}
\description{
Sample data for IBD16S data
}
\keyword{datasets}
