% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD16S_tax.R
\docType{data}
\name{IBD16S_tax}
\alias{IBD16S_tax}
\title{IBD16S_tax}
\format{An object of class \code{matrix} with 982 rows and 6 columns.}
\usage{
IBD16S_tax
}
\description{
Taxonomy data for IBD 16S data
}
\keyword{datasets}
