% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspi16S.R
\docType{data}
\name{momspi16S}
\alias{momspi16S}
\title{MOMS-PI 16S rRNA sequencing data}
\format{A phyloseq object with 7,665 taxa and 9,107 samples.
\subsection{OTU Table}{
    A counts matrix for the rRNA-seq results.
}
\subsection{Sample Data}{
   \describe{
      \item{sample_id}{Sample identifier}
      \item{subject_id}{Participant identifier}
      \item{sample_body_site}{Body site of the sample}
      \item{visit_number}{Visit number}
      \item{subject_gender}{Participant gender}
      \item{subject_race}{Participant race}
      \item{study_full_name}{Name of the study}
      \item{file}{Filename which the sample was taken from}
   }
}
\subsection{Taxonomy Table}{
    Taxonomy table for the taxa identified.
}}
\usage{
momspi16S
}
\description{
The Multi-Omic Microbiome Study-Pregnancy Initiative (MOMS-PI)
was funded by the NIH Roadmap HUman Microbiome Project to
understand the impact of the vaginal microbiome
on pregnancy and the fetal microbiome. This longitudinal study
contains samples from various body sites of 596 women. This phyloseq
object contains the 16S rRNA sequencing data that was collected.
Also contained are the taxa data and participant metadata.
}
\keyword{datasets}
