% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspi16S_mtx.R
\docType{data}
\name{momspi16S_mtx}
\alias{momspi16S_mtx}
\title{momspi16S_mtx}
\format{An object of class \code{matrix} with 7665 rows and 9107 columns.}
\usage{
momspi16S_mtx
}
\description{
16S rRNA OTU matrix for MOMS-PI data
}
\keyword{datasets}
