% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspi16S_samp.R
\docType{data}
\name{momspi16S_samp}
\alias{momspi16S_samp}
\title{momspi16S_samp}
\format{An object of class \code{data.frame} with 9107 rows and 9 columns.}
\usage{
momspi16S_samp
}
\description{
Sample data for MOMS-PI 16S data
}
\keyword{datasets}
