% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspi16S_tax.R
\docType{data}
\name{momspi16S_tax}
\alias{momspi16S_tax}
\title{momspi16S_tax}
\format{An object of class \code{matrix} with 7665 rows and 7 columns.}
\usage{
momspi16S_tax
}
\description{
Taxonomy data for MOMS-PI 16S data
}
\keyword{datasets}
