% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspiCyto_mtx.R
\docType{data}
\name{momspiCyto_mtx}
\alias{momspiCyto_mtx}
\title{momspiCyto_mtx}
\format{An object of class \code{matrix} with 29 rows and 1396 columns.}
\usage{
momspiCyto_mtx
}
\description{
Cytokines matrix for MOMS-PI data
}
\keyword{datasets}
