% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspiCyto_samp.R
\docType{data}
\name{momspiCyto_samp}
\alias{momspiCyto_samp}
\title{momspiCyto_samp}
\format{An object of class \code{data.frame} with 1396 rows and 13 columns.}
\usage{
momspiCyto_samp
}
\description{
Sample data for MOMS-PI cytokines data
}
\keyword{datasets}
