% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspiCytokines.R
\name{momspiCytokines}
\alias{momspiCytokines}
\title{Create MOMS-PI cytokines SummarizedExperiment object}
\format{A SummarizedExperiment object with 29 features
and 872 samples.
\subsection{colData}{
     \describe{
     \item{sample_id}{Sample identifier}
    \item{subject_id}{Participant identifier}
    \item{sample_body_site}{Body site of the sample}
    \item{visit_number}{Visit number}
    \item{subject_gender}{Participant gender}
    \item{subject_race}{Participant race}
    \item{study_full_name}{Name of the study}
    \item{file}{Filename which the sample was taken from}
     }
}}
\usage{
momspiCytokines()
}
\value{
A SummarizedExperiment object
}
\description{
The Multi-Omic Microbiome Study-Pregnancy Initiative (MOMS-PI)
was funded by the NIH Roadmap HUman Microbiome Project to
understand the impact of the vaginal microbiome
on pregnancy and the fetal microbiome. This longitudinal study
contains samples from three body sites of 116 women.
This summarizedExperiment
object contains the cytokine data that was collected and
the participant metadata.
}
\examples{
momspiCyto <- momspiCytokines()
}
