% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momspiMultiAssay.R
\name{momspiMultiAssay}
\alias{momspiMultiAssay}
\title{Make MOMS-PI MultiAssayExperiment}
\format{A MultiAssayExperiment object with a 16S rRNA matrix and Cytokine matrix
\subsection{16S}{
    A counts matrix for the 16S rRNA-seq results.
}
\subsection{cytokines}{
    A counts matrix for the cytokines results.
}
\subsection{colData}{
   \describe{
      \item{file_id}{File identifier}
      \item{md5}{md5 hash for the file}
      \item{size}{file size}
      \item{urls}{URL for the file}
      \item{sample_id}{Sample identifier}
      \item{file_name}{Filename which the sample was taken from}
      \item{subject_id}{Participant identifier}
      \item{sample_body_site}{Body site of the sample}
      \item{visit_number}{Visit number}
      \item{subject_gender}{Participant gender}
      \item{subject_race}{Participant race}
      \item{study_full_name}{Name of the study}
      \item{project_name}{Name of project}
   }
}}
\usage{
momspiMultiAssay()
}
\value{
A multiAssay Experiment object
}
\description{
Construct MultiAssayExperiment for MOMS-PI 16S rRNA
and cytokines data.
}
\examples{
momspiMA <- momspiMultiAssay()
}
