% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_table.R
\name{patient_table}
\alias{patient_table}
\title{Generate patient summary table for HMP2 sample data}
\usage{
patient_table(x)
}
\arguments{
\item{x}{A named list of phyloseq or SummarizedExperiment
objects.}
}
\value{
A knitr::kable table.
}
\description{
This function allows you to produce a summary table for
the HMP2Data data sets.
}
\examples{
patient_table(list(momspi16S = momspi16S(),
  momspiCytokines = momspiCytokines(),
  IBD16S = IBD16S(), T2D16S = T2D16S()))
}
