% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2D16S_mtx.R
\docType{data}
\name{T2D16S_mtx}
\alias{T2D16S_mtx}
\title{T2D16S_mtx}
\format{An object of class \code{matrix} with 10815 rows and 1319 columns.}
\usage{
T2D16S_mtx
}
\description{
T2D 16S rRNA matrix
}
\keyword{datasets}
