% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2D16S_samp.R
\docType{data}
\name{T2D16S_samp}
\alias{T2D16S_samp}
\title{T2D16S_samp}
\format{An object of class \code{data.frame} with 2208 rows and 13 columns.}
\usage{
T2D16S_samp
}
\description{
T2D 16S sample data
}
\keyword{datasets}
