% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2D16S_tax.R
\docType{data}
\name{T2D16S_tax}
\alias{T2D16S_tax}
\title{T2D16S_tax}
\format{An object of class \code{matrix} with 10815 rows and 7 columns.}
\usage{
T2D16S_tax
}
\description{
Taxonomy data for T2D 16S data
}
\keyword{datasets}
