% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_two.R
\name{table_two}
\alias{table_two}
\title{Generate demographic summary table for HMP2 sample data}
\usage{
table_two(x)
}
\arguments{
\item{x}{A named list of phyloseq or SummarizedExperiment
objects.}
}
\value{
A knitr::kable table.
}
\description{
This function allows you to produce a summary table for
the HMP2Data data sets.
}
\examples{
table_two(list(momspi16S = momspi16S(), momspiCytokines = momspiCytokines(), IBD16S = IBD16S(), T2D16S = T2D16S()))
}
