% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_table.R
\name{visit_table}
\alias{visit_table}
\title{Make table of visits}
\usage{
visit_table(x)
}
\arguments{
\item{x}{A named list of phyloseq or SummarizedExperiment
objects.}
}
\value{
A knitr::kable table.
}
\description{
This function allows you to produce a summary table for
the HMP2Data visits.
}
\examples{
visit_table(list(momspi16S = momspi16S(),
  momspiCytokines = momspiCytokines(),
  IBD16S = IBD16S(), T2D16S = T2D16S()))
}
